[symex] Move counterexample rendering into goto-symex/trace/
Third of seven commits partitioning src/goto-symex/
(docs/roadmap/goto-symex-partition-plan.md §2). goto_trace, build_goto_trace,
printf_formatter and the XML/HTML/JSON/SARIF renderers construct and render a
counterexample once one exists.

Pure move — the only content edits are #include paths. Note that trace/ does
not satisfy §3 rule 5: symex_printf includes printf_formatter.h and is called
from symex_assign and symex_other, so trace/ is reachable from symex_step. That
coupling is pre-existing; this commit neither adds nor removes it.
