+++ src/esbmc/parseoptions/command_line_options.cpp
@@ -20,8 +20,8 @@ extern "C"
 #include <esbmc/esbmc_parseoptions.h>
 #include <esbmc/globals.h>
 #include <goto-symex/goto_symex.h>
-#include <goto-symex/goto_trace.h>
-#include <goto-symex/sarif.h>
+#include <goto-symex/trace/goto_trace.h>
+#include <goto-symex/trace/sarif.h>
 #include <util/base/cwe_mapping.h>
 #include <solvers/smt/smt_result.h>
 #include <solvers/smtlib/smtlib_conv.h>
