+++ src/goto-symex/symex_target_equation.h
@@ -5,7 +5,7 @@
 #include <cstdio>
 #include <cstring>
 #include <goto-programs/goto_program.h>
-#include <goto-symex/goto_trace.h>
+#include <goto-symex/trace/goto_trace.h>
 #include <goto-symex/symex_target.h>
 #include <list>
 #include <map>
