+++ src/goto-symex/trace/goto_trace.h
