+++ src/goto-symex/trace/printf_formatter.h
