+++ src/goto-symex/trace/xml_goto_trace.h
@@ -1,7 +1,7 @@
 #ifndef CPROVER_GOTO_SYMEX_XML_GOTO_TRACE_H
 #define CPROVER_GOTO_SYMEX_XML_GOTO_TRACE_H
 
-#include <goto-symex/goto_trace.h>
+#include <goto-symex/trace/goto_trace.h>
 #include <util/base/xml.h>
 
 void convert(const namespacet &ns, const goto_tracet &goto_trace, xmlt &xml);
