+++ src/goto-symex/witness/witnesses.h
@@ -5,7 +5,7 @@
 #include <util/symtab/namespace.h>
 #include <irep2/irep2.h>
 #include <langapi/language_util.h>
-#include <goto-symex/goto_trace.h>
+#include <goto-symex/trace/goto_trace.h>
 #include <goto-symex/symex_target_equation.h>
 #include <string>
 #include <regex>
